requirements.txt
numpy
pandas
scipy
matplotlib
seaborn
opencv-python
moviepy
pydub
plotly
pillow
nbformat>=4.2.0
brokenaxes
pytk